+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayer.java
@@ -436,6 +436,7 @@ public interface ExoPlayer extends Player {
      * @param timeoutMs The time limit in milliseconds.
      */
     public Builder experimentalSetForegroundModeTimeoutMs(long timeoutMs) {
+      Assertions.checkState(!buildCalled);
       setForegroundModeTimeoutMs = timeoutMs;
       return this;
     }